refactor(foldable): move fold into its own interface
