+++ src/fold/index.ts
@@ -1,0 +1,1 @@
+export { Foldable, Folding } from './foldable';
