+++ src/railway/index.ts
@@ -1,2 +1,2 @@
-export type { Folding, Railway } from './railway';
+export type { Railway } from './railway';
 export { AsyncRailway } from './async-railway';
