+++ src/railway/railway.test.ts
@@ -4,11 +4,6 @@ import { Either } from '../either';
 import { Option } from '../option';
 
 describe('Railway', () => {
-  it.each(foldTestCases)('$type should handle fold operation correctly', ({ railway, folding, expected }) => {
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-expect-error
-    expect(railway.fold(folding)).toEqual(expected);
-  });
   it.each(andThenTestCases)('$type should handle andThen operation correctly', ({ railway, operation, expected }) => {
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-expect-error
@@ -29,63 +24,6 @@ describe('Railway', () => {
     }
   );
 });
-
-const foldTestCases = [
-  {
-    type: 'Either Right',
-    railway: Either.right<number, number>(2),
-    folding: {
-      ifRight: (x: number) => x * 2,
-      ifLeft: (_: number) => 2,
-    },
-    expected: 4,
-  },
-  {
-    type: 'Either Left',
-    railway: Either.left<number, number>(2),
-    folding: {
-      ifRight: (_: number) => 2,
-      ifLeft: (x: number) => x * 2,
-    },
-    expected: 4,
-  },
-  {
-    type: 'Try Success',
-    railway: Try.success(2),
-    folding: {
-      ifSuccess: (x: number) => x * 2,
-      ifFailure: (_: Error) => 2,
-    },
-    expected: 4,
-  },
-  {
-    type: 'Try Failure',
-    railway: Try.failure<number>(new Error('fail')),
-    folding: {
-      ifSuccess: (_: number) => 2,
-      ifFailure: (e: Error) => e.message,
-    },
-    expected: 'fail',
-  },
-  {
-    type: 'Option Some',
-    railway: Option.of(2),
-    folding: {
-      ifSome: (x: number) => x * 2,
-      ifNone: () => 2,
-    },
-    expected: 4,
-  },
-  {
-    type: 'Option None',
-    railway: Option.of<number>(undefined),
-    folding: {
-      ifSome: (_: number) => 2,
-      ifNone: (x: undefined) => x,
-    },
-    expected: undefined,
-  },
-];
 
 const andThenTestCases = [
   {
